gptel: No advice for tool spec argument preprocessing (#794)
* gptel.el (gptel--preprocess-tool-args, gptel--make-tool,
gptel-tool, gptel--make-tool-internal):
To work around inconsistent advice evaluation for cl-struct
constructors, separate tool construction into two functions:
`gptel--make-tool-internal' is the cl-struct constructor, now
unadvised.  `gptel--make-tool' runs the constructor after
preprocessing the tool spec to change :types to strings (and
possibly other accommodations in the future) the tool spec to
change :types to strings (and possibly other accommodations in the
future).  Leave `gptel-make-tool' unchanged.

The problem of inconsistent advice application to cl-struct
constructors (which have a compiler macro attached) can be seen
in the following example.  Place the following in an elisp file
and native-compile and load it:

  ;; -*- lexical-binding: t; -*-
  (cl-defstruct teststruct foo)

  (define-advice make-teststruct (:filter-args (rest) modify-args)
    (plist-put rest :foo "changed")
    rest)

Now running (make-teststruct :foo "original") will cause the
struct to have its :foo key set to either "original" or "changed",
depending on whether the advice ran.  Whether the advice runs
appears to depend on the environment and the method of evaluation.

- The advice runs when evaluated in IELM, via pp-eval-last-sexp or
pp-eval-expression.

- The advice doesn't run when evaluated via eval-last-sexp,
eval-defun or eval-buffer.

This requires further investigation.