gptel-org: Don't include tags in gptel-org-set-topic (#801)
gptel-org.el (gptel-org-set-topic): The function `org-get-heading`
returns the heading text along with any tags if the heading has
tags. In contrast, the `ITEM` property gives the actual heading
text of the entry and should be used for that purpose. When
available, the function uses the org-element cache to improve
efficiency.