gptel: Ignore non-streaming blank responses
If the non-streaming response is all whitespace, do not call the
gptel-request callback.  This causes `gptel-response-separator' to
be inserted without any text preceding it. (#819)

Note that prior to this commit, we string-trimmed responses
already, so whitespace at the beginning and ends of responses was
not respected.  Blank responses were thus passed as empty strings
to the callbacks.  As of this commit, blank responses are not sent
to the callback at all.

If it turns out that blank responses are significant for special
tasks/gptel-request callbacks, we can move this check to the
default callback, `gptel--insert-response' and call the callback
with blank responses.

* gptel.el (gptel--url-parse-response): Make the change.

* gptel-curl.el (gptel-curl--parse-response): Make the change.