WIP feature-augment-buffer
Note: A number of things are broken as of this commit:
- gptel-context commands
- gptel-rewrite
- directives that are functions or lists
- Passing a list of prompts to gptel-request

This commit is for testing augmentation with standard gptel use
only.  Here is how to do that.

A synchronous augmentor:

- is a function placed in gptel-augment-handler-functions.

- is called with one argument, the fsm of the request being
augmented.  (You do not need to use this fsm except in rare
cases.)

- When it is called, the current buffer is the one whose contents
will be parsed as the conversation/messages array.  Let's call it
the "prep buffer".  The cursor is at (point-max) in this buffer.

- You are free to edit this buffer however you wish, such as by
adding RAG results, or changing request parameters based on the
presence of an @agent string in the prompt.  The commands
`gptel-beginning-of-response' and `gptel-end-of-response' might be
helpful.

- You can also search for the text-property `gptel' -- regions
where it is non-nil are part of an LLM response or a tool call +
result.

- To change the model, backend, system message or tools you can
set the standard gptel user options (gptel-model,
gptel-backend, gptel--system-message, gptel-tools etc) with setq.
These will be set buffer-locally in this prep buffer.

- When you are done editing this buffer or gptel settings, there
is nothing left to do.  You can simply return.

An asynchronous augmentor is similar, except for these two
differences:

- It is called with two arguments, a callback and the fsm of the
request being augmented.

- When you are done editing the prep buffer or gptel settings, you
have to call the callback.  The callback is called with NO
arguments, it is a thunk.

That's it.