gptel: Test new prompt-list format for gptel-request
Pre-existing behavior: `gptel-request' can be supplied a list of
strings for its PROMPT argument, representing a
conversation (messages array) with alternating roles (user/llm).

However if the conversation involves tool calls and results, this
format is insufficient.  Following the discussion in #811, add
support for a new format with explicit roles.  PROMPT may now be
specified as

 ((prompt   "User message 1")
  (response "LLM reply 1")
  (prompt   "User message 2")
  (prompt   "User message 3")
  (response "LLM reply 2")
  (tool     :name "some_function"
            :args (:query "test" :limit 10)
            :id "call_abcd"                 ; OPTIONAL
            :result "tool result")
  (response "LLM reply 3"))

NOTE: This change is experimental and otherwise undocumented, and
the format is expected to change in the future.

The pre-existing format of a list of strings will continue to be
supported.

* gptel.el (gptel-request): Add TODO about documenting new
prompt-list input format.

* gptel-openai.el (gptel--parse-list): Make the change.

* gptel-ollama.el (gptel--parse-list): Make the change.

* gptel-gemini.el (gptel--parse-list): Make the change.

* gptel-anthropic.el (gptel--parse-list): Make the change.