gptel: Implement presets (#542)
One of the main points of friction with using gptel is the need to
repeatedly change several individual options like the backend,
model, system message, included tools etc, for each task or in
each buffer.  Presets allow combinations of options to be
pre-specified and applied together, making it feasible to switch
rapidly between different kinds of LLM tasks.

* gptel.el (gptel-make-preset, gptel-get-preset):
(gptel--apply-preset): Implement option presets for gptel.
`gptel-make-preset' is the entry point for defining presets.  An
option preset is a plist of settings that can be retrieved by name
using `gptel-get-preset', and applied via `gptel--apply-preset'.

`gptel--apply-preset' is an internal function because the primary
method of applying presets is intended to be via gptel's transient
menu.  This will be implemented next.