gptel-openai: Prevent null tool-calls from causing error
Ensure that tool-calls are processed only if they are not null. This
issue was discovered when using LiteLLM, and is possibly an
implementation bug there. At least differs from the real OpenAI.com
API, which doesn't have this issue. However, for gptel to be usable
with LiteLLM, this check is necessary.