C-y can be used to add the current kill to the context
This saves the user from having to first create a new buffer, yank the current
kill into that buffer with C-y, and then add that buffer as context. With this
change, they can just type C-y in the transient gptel-send buffer and the
temporary is created as needed.

If there are already kills that have been yanked, the new kills will be added
to the context. In order to reset this buffer, use C-u C-y when yanking from
the transient menu.