gptel-transient: Preset switching via completing-read
* gptel-transient.el (gptel--preset-variable, gptel--preset,
gptel--preset-mismatch-p, transient-infix-set,
transient-format-value, gptel--infix-preset): gptel's transient
menu now shows the active preset correctly, and indicates if the
preset's settings have been manually modified afterwards.

Introduce `gptel--preset', an internal variable to hold the last
applied preset name.  This is for persistent display of the active
preset in the transient menu.  The corresponding infix is now of
class `gptel--preset-variable', which derives from
`gptel-lisp-variable'`, which derives from `gptel-lisp-variable'.

`gptel--preset-mismatch-p' checks if the preset is in effect by
verifying the relevant state variables.  Despite a hand-rolled
loop, this function can be slow if there is a lot of state to
check.  This feature remains experimental.