gptel: Add command to save a preset
* gptel.el (gptel--save-preset): Add a command to save gptel's
current settings as a preset.  In addition to registering a
preset, lisp code to do the same is copied to the kill-ring.  This
is for the benefit of users who might want to record it in their
configurations.

The new preset covers a few common gptel settings and does not
attempt to cover everything.