gptel-transient: Add a preset menu, remove infix
* gptel-transient.el:
(gptel--format-preset-string, gptel--preset): Use a new preset
menu instead of an infix to manage presets.  The design is similar
to that of `gptel-system-prompt', and includes a command to save
the current settings to a preset.  Additionally, the "Request
Parameters" heading now always shows an "(@preset)" string, since
a new preset may be created at any time, even when the registry is
nil.

(gptel-menu): Add new preset.

(gptel--infix-preset, gptel--preset-variable, gptel--preset):
Remove the preset-variable class and methods, as they are no
longer required.