gptel: Add prompt-transform functions
* gptel.el:
(gptel-prompt-filter-hook, gptel-prompt-transform-functions):
Introduce a hook to hold transformations to apply to the prompt.

This will be used to hold functions for context injection, Org
property stripping and so on, as well as just-in-time preset
application and any RAG plugins.  This hook will be passed
explicitly to `gptel-request' as an argument, since
`gptel-request' is supposed to be a simple function by default,
with behavior that depends as little on global variables as
possible.

Obsolete `gptel-prompt-filter-hook', since it is superseded by
`gptel-prompt-transform-functions'.

This hook currently does nothing, and will be added to the
`gptel-request' pipeline next.

(gptel--with-buffer-copy, gptel--create-prompt-buffer): Remove
gptel-prompt-filter-hook from the prompt-buffer creation pipeline.

* gptel-org.el (gptel-org--create-prompt): Remove
gptel-prompt-filter-hook from the prompt-buffer creation pipeline.