gptel: Switch to prompt-buffer centric parsing
Previously, gptel's prompt construction worked as follows:

1. If the prompt is nil or a string, copy the text to be sent to a
temporary buffer, and use it to create the messages array.  If the
prompt is a list of strings, create the messages array from it
directly (with no intermediate buffer).

2. Feed this array along with the specified or environmental
parameters like the backend, model, system message and other
parameters to `gptel--request-data', which produces the full
payload.

gptel is now inverting this flow:

1. Whatever the prompt, copy it to a temporary buffer (the "prompt
buffer") using `gptel--create-prompt-buffer' or directly with
`gptel--with-buffer-copy'.  Use the new function
`gptel--parse-list-and-insert' to insert a list-of-strings prompt
into the prompt buffer.

2. Also copy the environment to this temporary buffer as
buffer-local variables.

3. Apply any augmentations/transformations/context injections to
this prompt buffer.  This can involve changing the buffer contents
and/or changing the environment, i.e. buffer-local variables such
as the model, backend or system prompt.  The ability to do this
is the main reason for this inversion.  One example of this is
the upcoming just-in-time preset feature.

NOTE: transformations are not yet implemented.

4. Construct the full payload using `gptel--request-data' inside
the prompt buffer.

* gptel.el:
(gptel--parse-directive): Add FIXME for later
(gptel--parse-list-and-insert): Inserts prompts, a list of
strings, into the current buffer with user/llm/tool roles applied.
(gptel-request, gptel--realize-query): Split the work of
`gptel-request' into two functions.  `gptel-request' now handles
only the creation of the prompt buffer.  `gptel--realize-query' runs
in the prompt buffer and does the actual payload realization.
(gptel--create-prompt): Remove function
(gptel--create-prompt-buffer): Renamed from
`gptel--create-prompt', now returns the prompt buffer instead of a
messages array.  No longer handles context injection.

* gptel-org.el (gptel-org--create-prompt): Remove function
(gptel-org--create-prompt-buffer): Renamed from
`gptel-org--create-prompt', now returns a buffer instead of a
messages array.