gptel: Apply prompt-transform-functions in gptel-request
Use `gptel-prompt-transform-functions' in `gptel-request',
enabling a prompt transformation/augmentation pipeline.  This
pipeline will be used for just-in-time presets, adding gptel's context
to the request and other operations that are critical to
`gptel-send'.

`gptel-prompt-transform-functions' is not automatically applied to
`gptel-request' calls, it must be passed explicitly via the
`:transforms' argument.  This is so that gptel-request stays
closer to a 'pure' function, for use as a library.

`gptel-send' passes the transforms explicitly to `gptel-request'.

* gptel.el (gptel--create-prompt-buffer, gptel--create-prompt):
New function `gptel--create-prompt-buffer', used to create a
prompt buffer with the appropriate text to be
sent (pre-transform).  `gptel--create-prompt' is no longer used
but retained for compatibility and testing.

(gptel-request): Add the `:transforms' argument and docstring.

Run `gptel-prompt-transform-functions' here.  In short, create the
prompt construction buffer with `gptel--create-prompt-buffer',
then run through the transforms.  The preset transform is applied
first, since it can change the value of
`gptel-prompt-transform-functions'.  The rest are applied via a
wrapped hook that distinguishes between sync and async transforms
via function arity, and uses a "gather" method to track
completion.

TODO: Async transforms end up running in parallel, so there is a
race condition involving buffer mutation when there is more than
one.  This will be fixed in the future.

* gptel-transient (gptel--suffix-send): Add `:transforms' to the
`gptel-request' call.

* gptel-rewrite (gptel--suffix-rewrite): Add `:transforms' to the
`gptel-request' call.