gptel: Fix dispatch on prompt types in gptel--parse-list
`gptel-request' accepts two kinds of prompt lists as arguments --
a list of strings and a list of lists/cons cells.
(The second kind is currently experimental and undocumented.)  In
`gptel--parse-list', we distinguish between them by checking if
the car of the prompt is a string.  But the car can also be nil,
and this check fails.  Fix by checking if the car of the prompts
list is a cons instead.

* gptel-anthropic.el (gptel--parse-list): Make the change.
* gptel-gemini.el (gptel--parse-list): Make the change.
* gptel-ollama.el (gptel--parse-list): Make the change.
* gptel-openai.el (gptel--parse-list): Make the change.