gptel: Turn context addition into a transform step
Previously, adding context (from gptel-context) to the query
payload was a bespoke step, and tied into `gptel-request'.  It was
carried out by mutating the messages array after buffer parsing.

Now that the `gptel-request' pipeline has been inverted, context
addition is carried out in the prompt construction buffer, and is
handled as a regular transform function in
`gptel-prompt-transform-functions'.  As a result, the context
addition step can now be asynchronous, since the user might want
to truncate or augment it depending on its size and content.  By
default, context addition works by inserting into the prompt
construction buffer, or by modifying the system message in the
buffer.

The only exception to this is handling media files in the
context -- there is no universal buffer-oriented method for doing
this, so `gptel--wrap-user-prompt' is still used for attaching
media from `gptel-context--alist'.
TODO: This will be fixed in the future.

This deprecates some aspects of the gptel-context API, including `gptel-context-wrap-function'.

* gptel-context.el (gptel-context-wrap-function): Obsolete.  This
is no longer used.  Customize `gptel-context-string-function'
instead or advise `gptel-context--wrap-in-buffer' if that's
insufficient.
(gptel-context--wrap): This function serves the same purpose as
before, but now works by mutating the prompt construction buffer.
It takes a callback and can be asynchronous.
(gptel-context--wrap-in-buffer): New function to add the context
string to the prompt construction buffer, etiher as text or by
modifying the buffer's system message.
(gptel-context--wrap-default):  Remove, as this is covered by
`gptel-context--wrap-in-buffer'.

* gptel.el (gptel--transform-add-context): Augmentor for adding
context to the request, which calls gptel-context- functions.
This will be added to `gptel-prompt-transform-functions'
next.  (Context addition is currently non-functional in gptel.)
(gptel--realize-query): Move context addition step from here back
to the transforms/augmentation step.  Handling media files in the
context is an exception and it still happens here.
(gptel--wrap-user-prompt-maybe): Remove, no longer needed.