gptel: Set default for gptel-prompt-transform-functions
* gptel.el (gptel-prompt-transform-functions): Set the default
transformers/augmentors.  They (i) apply any preset included at the
beginning of the prompt (via `@preset`) and (ii) add gptel's
context to the prompt buffer.

Note that `gptel-prompt-transform-functions' (or a list containing
augmentors) needs to be passed explicitly to `gptel-request' (via
the `:transforms' argument) for it to take effect.  `gptel-send'
does this.  It is not read from the environment as this can have
unexpected side-effects when using `gptel-request' as a library.