gptel: Send links pointing to plaintext files (#475, #481)
When `gptel-track-media' is enabled, follow links in gptel
buffers (Org or Markdown) that point to plaintext files.

The file text is included in place of the link when sending
requests.  This provides an easy way to provide buffer-local
context in chat buffers.

Previously, this option only controlled whether links to binary
files (images/documents) of supported mime-types were included
with the request.

NOTE: `gptel-track-media' will be renamed to reflect this
expansion in scope next.

* gptel.el (gptel--file-binary-p, gptel--insert-file-string): Move
helper functions from gptel-context, as they are needed for
inserting linked files.

(gptel--parse-media-links): Update Markdown implementation to
handle links to plaintext files.

* gptel-org.el (gptel--parse-media-links): Update Org mode
implementation to handle links to plaintext files.

* gptel-context.el (gptel--file-binary-p)
(gptel-context--insert-file-string): Move to gptel.el and rename.
(gptel-context--string): Adjust for renaming.

* gptel-anthropic.el (gptel--anthropic-parse-multipart): Update
to insert the contents of plaintext files.
* gptel-gemini.el (gptel--gemini-parse-multipart): Ditto.
* gptel-ollama.el (gptel--ollama-parse-multipart): Ditto.
* gptel-openai.el (gptel--openai-parse-multipart): Ditto.