gptel: Make track-media always available
Since links to plain-text files can now always be included in chat
buffers (irrespective of the model's media capabilities), this
option should always be available in the transient menu, and
whether we scan for links should depend only on
`gptel-track-response', and not on the model capabilities.  Make
these adjustments to the buffer parsing call chain.

The "Sending media" option still does not show up in the header
unless the model is media-capable.  This might be something to
work on.

* gptel-transient.el (gptel-menu): Allow `gptel-track-media' to
always be toggled in chat buffers.

* gptel.el (gptel--insert-file-string): Fix bug around file
insertion -- `insert-file-contents' does not move point, so we
must move it ourselves.

* gptel-anthropic.el (gptel--parse-buffer): Fix buffer parsing
code path as described.
* gptel-gemini.el (gptel--parse-buffer): Ditto.
* gptel-ollama.el (gptel--parse-buffer): Ditto.
* gptel-openai.el (gptel--parse-buffer): Ditto.