gptel: Change preset :after function to :post, add :pre
* gptel.el (gptel-make-preset): Change `:after' to `:post' as a
recognized keyword in gptel presets.  It's value is expected to be
a function that runs after the preset is applied.

Also add a `:pre' key for a function to run before a preset is
applied.

(gptel--apply-preset): Run the `:pre' and `:post' functions before
and after applying a preset, respectively.

* gptel-transient.el (gptel--preset-mismatch-p): Adjust for
`:pre' and `:post'.