gptel-transient: Redesign tools menu with two-column layout
Redesign the gptel-tools menu to use a two-column layout for tools
and categories.  The primary reason for the change is that the
previous tools menu did not scale to large numbers of tools (>
50), a number very easy to get to with MCP-provided tools.  Now
only one category of tools is shown at a time.  As a bonus, the
new menu requires half as many keystrokes as before to enable
tools or toggle tool categories.

This redesign is inspired by the wonderful proof-of-concept
demonstration by @ahmed-shariff in #653.

The implementation is necessarily more complex, with a design
unlike most Transient prefixes.  Some notes:

Normally a transient prefix exports its value via transient-args,
to be consumed by suffixes, where these args are determined by the
state of the menu at the time of export.  The gptel-tools menu is
dynamic and needs to store tool selections that may not be visible
in the meny any more, so we cannot use the transient-args as is.

We can not (should not?) control the value of the prefix directly,
so we instead use the scope (a secondary value) of the prefix to
maintain the history of selections.  When running a suffix, we
gather tool selections from the scope.  The scope is also used as
a message channel for connecting the category group and the tool
list group for that category.

* gptel-transient.el (transient-infix-set, transient-infix-read)
(transient-format-value, gptel--switch-category, gptel--switch)
(gptel-tools): Redesign menu as described.  The `gptel--switch'
and `gptel--switch-category' class methods are modified.

* gptel-integrations.el (gptel--suffix-mcp-connect,
gptel--suffix-mcp-disconnect): Update the state of the gptel-tools
menu correctly and carefully after adding or removing tools from
MCP servers.

* NEWS: Update with mention of new tools menu.