gptel: Add gptel--request-params for extra request options
* gptel.el (gptel--request-params)
(gptel--with-buffer-copy-internal): Add new variable for
specifying global/buffer-local/oneshot request-params.

Important: It is very easy to break gptel with stale/incompatible
request-params.  As such, request-params are best specified as
backend or model properties.  If used, `gptel--request-params' is
best set via a preset, and it's the user's responsibility to unset
this when switching presets.  For this reason it is marked as an
internal variable.  Really, it is only here because of the unique
configuration requirements of John Wiegley.

* gptel-anthropic.el (gptel--request-data): Use it in request
payload.
* gptel-gemini.el (gptel--request-data): Use it in request
payload.
* gptel-ollama.el (gptel--request-data): Use it in request
payload.
* gptel-openai.el (gptel--request-data): Use it in request
payload.
* gptel-kagi.el:  Update TODO about request-params support.