gptel-org: Set org-complex-heading-regexp in prompt buffer
* gptel-org.el (gptel-org--create-prompt-buffer): Copy
`org-complex-heading-regexp' from the request buffer to the prompt
construction buffer.  This is required for `org-element-context'
to run when parsing links. (#947)

Note that the more robust way to do this is to run
`org-set-regexps-and-options' instead.  But I'm not happy with the
performance of this function, it does a surprising amount of
consing for something that ostensibly just sets a few variables.
We can move to this mitigation if other problems arise as a result
of not calling `org-mode' in the prompt construction buffer.