gptel-curl: Fix text handling around end of reasoning block
* gptel-curl.el (gptel-curl--stream-filter): At the end of a
reasoning block, the response string is non-nil but we end up
treating it as reasoning because of the way <think>..</think>n
blocks are handled.  This causes the first chunk of the response
to be painted as reasoning text.

To fix this, we need a way to distinguish between the two kinds of
reasoning blocks (separate reasoning field vs embedded as
<think>..</think> blocks in the response).  Use the fact that when
reasoning text obtained from a reasoning_content (etc) field in
the response, :reasoning-block is set to t to indicate the
response stream has ended.