gptel: Add copy Curl command feature to dry-run
* gptel.el (gptel--inspect-query, gptel--continue-query): Allow
the Curl command for a request to be copied from the dry-run
inspection buffer.  This is bound to C-c C-w and indicated in the
header-line.

Note that the payload that is sent is determined by this buffer's
contents, which are editable.  (The same is true when you continue
the command with C-c C-c.)