gptel: Add JSON output for OpenAI, Anthropic, Gemini, Ollama
Activate support for structured outputs to the OpenAI, Anthropic,
Gemini and Ollama backends.  Some other backends like llama-cpp
and Perplexity also work because they have OpenAI-compatible APIs.

Note: There are many other providers that purport to provide OpenAI
compatible APIs that do not support structured outputs.  This
includes Groq, Openrouter, Deepseek and others.

The approach is identical for all backends: if we find a schema in
`gptel--schema', we process it to a backend appropriate format and
request JSON output in the payload.  `gptel--schema' is intended
to be let-bound around calls to `gptel--request-data'.  A less
stateful solution for `gptel--request-data' is planned for the
future.

* gptel-anthropic.el (gptel--request-data, gptel--parse-schema):
The Anthropic API uses tool calls to (reliably) generate JSON
output.  Define an ersatz tool and check for JSON output as a
special case in `gptel--handle-tool-use'.  This approach has some
disadvantages: we cannot use both tools and require JSON output
in the same request.  Them's the breaks.

In all other cases, we use a backend-appropriate flag in the
payload to require JSON output.

* gptel-gemini.el (gptel--request-data, gptel--parse-schema):
* gptel-openai.el (gptel-tools, gptel--parse-tools)
gptel--request-data, gptel--parse-schema):
* gptel-ollama.el (gptel--request-data):

* NEWS: Mention structured output support, the previously added
command to copy the Curl command for a request from the dry-run
buffer and Open WebUI support.