gptel-openai: Don't parse tools named "null" (#935) (#951)
* gptel-openai.el (gptel-curl--parse-stream): Temporarily ignore
tools named "null" (the string, not the JSON type) when parsing
tools during streaming.  This is for compatibility with the
litellm API, but this problem is on their side.

NOTE: This change will be reverted without warning at the time
of the next gptel release.  (If litellm hasn't fixed it by then,
your usage might break.)

Co-authored-by: karthink <karthikchikmagalur@gmail.com>