gptel-anthropic: Update cache control to include ttl "1h"
* README.org ((Optional) Interim support for Claude 3.7 Sonnet):
Simplify the setup for Claude-thinking.  None of the beta headers
are required any longer.

* gptel-anthropic.el (gptel--request-data, gptel--parse-list,
gptel--parse-buffer): Increase default cache ttl for system and
messages to 1 hour.  Tools cache remains at a ttl of 5 minutes.
It's not clear from the documentation if 1 hour caches are any
more expensive than 5 minute caches, but on the whole this should
reduce the costs of conversations when using `gptel-cache' with
Anthropic (only).

(gptel-make-anthropic): Remove obsolete headers, add
extended-cache-ttl support.