gptel: Enforce uniform convention for calling  presets
* gptel.el (gptel--apply-preset, gptel--preset-syms,
gptel-with-preset, gptel--transform-apply-preset): Ensure that the
preset-related functions `gptel--apply-preset',
`gptel--preset-syms' and the macro `gptel-with-preset' all refer
to presets the same way: A preset can be passed to these by name
or as a (possibly anonymous) preset spec, a plist.  Previously
they each had different conventions for their preset argument.

All of these are now valid:

  ;; Refer to preset by name
  (gptel--apply-preset 'websearch)
  (gptel--preset-syms  'websearch)
  (gptel-with-preset   'websearch)

  ;; Refer to preset by full spec
  (gptel--apply-preset '(:tools ... :stream nil))
  (gptel--preset-syms  '(:tools ... :stream nil))
  (gptel-with-preset   '(:tools ... :stream nil))

Note that `gptel-with-preset' now evaluates its PRESET argument
and it must be quoted as above.

* gptel-transient.el (gptel--preset): Adjust for preset API
changes.