gptel: Fix async prompt transform + gptel-send issues
* gptel.el (gptel--realize-query): Fix call to
`gptel-augment-post-modify-hook', which does not and has never
existed.  (How did it get there?)

(gptel-send): When there is an async prompt transform,
gptel-send's message runs before the `:backend' field has been set
in INFO.  Fix by falling back to the value of `gptel-backend' in
`gptel-send'. This might be wrong, depending on what the prompt
transform does, but at least it doesn't abort the process.