gptel: Make response tracking more robust
* gptel.el (gptel--inherit-stickiness, gptel-mode): Finally bite
the bullet and add an after-change-function to gptel-mode to track
response regions more robustly.  Now any change to a chat buffer
in an LLM response region will be marked as a response, mostly
resolving a number of problems and issues (#321, #343 and others).

This change is experimental, and will be reverted if there are
performance issues.  Preliminary testing suggests that it does not
cause any.

It also affects only chat buffers, so the
problem persists if you (i) use multi-turn requests (ii) via
gptel-send (iii) in non-chat buffers.  (gptel-rewrite and other
ways of using gptel don't suffer from the response tracking
issue.)