gptel: Allow JSON schemas to be specified via shorthand
* gptel.el (gptel--dispatch-schema-type, gptel-request): Allow
gptel-request's SCHEMA argument to be specified via shorthand
notation (for simple objects or arrays of objects) in addition to
the serialized JSON or a plist.  Most often, the required JSON
structure is simple and specifying the full schema is
annoying and error-prone.

See the documentation of `gptel--dispatch-schema-type' for
details on the shorthand format(s).

* NEWS (New features and UI changes): Update to mention new
capabilities.

* test: Update with unit tests for `gptel--dispatch-schema-type'.