gptel-rewrite: Start rewrite without setting up transient
To reduce the friction of using gptel-rewrite, don't invoke the
transient menu unless the user asks for it.  Instead, read a
rewrite instruction and start the rewrite immediately.

If no region is selected and there are pending rewrites, display
the rewrite menu.  This behavior is unchanged from before.

* gptel-rewrite.el (gptel--rewrite-read-message, gptel-rewrite)
(gptel--infix-rewrite-extra): New function
`gptel--rewrite-read-message' to read the rewrite instruction and
start the rewrite automatically if required.

* gptel-transient.el (gptel--read-with-prefix-help,
gptel--infix-add-directive): Minor changes to the usage of the
help string `gptel--read-with-prefix-help'.

* NEWS: Document `gptel-rewrite' behavior change.