gptel-rewrite: Make minibuffer reads editable in a buffer
When gptel asks you to type in a prompt or instruction in the
minibuffer, there is a hidden feature: you can switch to editing
it in a dedicated buffer with C-c C-e.  In Emacs 29.1 and up, the
`string-edit' package provides this out of the box, but we roll
our own since gptel supports Emacs 27.2+ and `string-edit' is not
included in the compat library.

This feature is undocumented as there is nowhere to display this
without overwhelming the user.  This commit documents changes to
this "directive-editor" feature.

* gptel-transient.el (gptel--edit-directive, gptel--suffix-send):
Add the ability to switch to a directive-editor buffer when typing
in the prompt in the minibuffer, with C-c C-e (#933).  This
feature is already available in other contexts.

Make the directive editor callback behavior distinguish between a
successful and aborted edit.  This is required for specifying
rewrite messages via the directive editor.

(gptel--read-crowdsourced-prompt, gptel--suffix-system-message):
The crowdsourced system prompt editor and the gptel directive
editor now take you back to gptel-menu when you're done. (#965)

(gptel--infix-add-directive): Add TODO about allowing composition
via the directive editor (C-c C-e) for this infix.  This feature
does not exist yet because of issues with communicating the
additional instruction as a scope argument to gptel-menu.  Unlike
in the other cases, there is no elisp symbol to be set here, the
result is an infix value that needs to be part of the prefix's
scope.

* gptel-rewrite.el (gptel--rewrite-read-message,
gptel--suffix-rewrite-directive): Confirming a rewrite message in
the directive editor now starts the rewrite, and canceling it
aborts the rewrite.