gptel-transient: Include buffer region when prompting from minibuffer
* gptel-transient.el (gptel--read-with-prefix,
gptel--read-minibuffer-prompt): When reading a prompt from the
minibuffer (the "m" option in gptel-menu), optionally include the
buffer region if it is active.  Region inclusion can be toggled on
demand from the minibuffer prompt via an indicated keybind.  As
before, the prompt can be edited in a separate buffer with C-c
C-e, an non-advertised option. (#969)

(gptel--suffix-send): Make appropriate change.