gptel-openai: Handle reasoning_content in gptel-openai (#1028)
Look for "reasoning_content" fields in the OpenAI-API response
parsers.  Previously we only looked for "reasoning" fields,
delegating "reasoning_content" handling to the separate
gptel-deepseek backend.

It turns out that OpenAI-compatible APIs are now split on this:

- "reasoning": Used by Openrouter and some others
- "reasoning_content": Used by vLLM, Llama.cpp, sglang, Deepseek
and others

A separate Deepseek backend is still required as it has extra
requirements for the messages array format, see the implementation
of `gptel--parse-buffer' for Deepseek.

* gptel-openai.el (gptel-curl--parse-stream,
gptel--parse-response): Handle "reasoning_content" field.

* gptel-openai-extras.el (gptel-curl--parse-stream,
gptel--parse-response): Remove methods specialized to Deepseek for
handling "reasoning_content", as the underlying OpenAI
implementation will handle this now.